chore: cleaned plonk bn254, removed old version
