+++ internal/backend/bn254/plonk/prove.go
@@ -31,36 +31,34 @@ import (
 	fiatshamir "github.com/consensys/gnark-crypto/fiat-shamir"
 )
 
-// ProofRaw PLONK proofs, consisting of opening proofs
-type ProofRaw struct {
+type Proof struct {
 
-	// Claimed Values of L,R,O,Z,H at zeta (where H split in H1+X^m*H2+X^2m*H3, where deg(Hi)=m-1)
-	LROZH [7]fr.Element
+	// Commitments to the solution vectors
+	LRO [3]polynomial.Digest
 
-	// Claimed vales of Z(zX) at zeta
-	ZShift fr.Element
+	// Commitment to Z, the permutation polynomial
+	Z polynomial.Digest
 
-	// Commitments to L,R,O,Z,H
-	CommitmentsLROZH [7]polynomial.Digest
+	// Commitments to h1, h2, h3 such that h = h1 + Xh2 + X**2h3 is the quotient polynomial
+	H [3]polynomial.Digest
 
-	// batch opening proofs for L,R,O,H,Z at zeta
-	BatchOpenings polynomial.BatchOpeningProofSinglePoint
+	// Batch opening proof of h1 + zeta*h2 + zeta**2h3, linearizedPolynomial, l, r, o, s1, s2
+	BatchedProof polynomial.BatchOpeningProofSinglePoint
 
-	// opening proof for Z at z*zeta
-	OpeningZShift polynomial.OpeningProof
+	// Opening proof of Z at zeta*mu
+	ZShiftedOpening polynomial.OpeningProof
 }
 
-// ComputeLRO extracts the solution l, r, o, and returns it in lagrange form.
+// ComputeLROBis extracts the solution l, r, o, and returns it in lagrange form.
 // solution = [ public | secret | internal ]
-func ComputeLRO(spr *cs.SparseR1CS, publicData *PublicRaw, solution []fr.Element) (bn254.Polynomial, bn254.Polynomial, bn254.Polynomial, bn254.Polynomial) {
+func ComputeLROBis(spr *cs.SparseR1CS, pk *ProvingKey, solution []fr.Element) (bn254.Polynomial, bn254.Polynomial, bn254.Polynomial) {
 
-	s := int(publicData.DomainNum.Cardinality)
+	s := int(pk.DomainNum.Cardinality)
 
-	var l, r, o, partialL bn254.Polynomial
+	var l, r, o bn254.Polynomial
 	l = make([]fr.Element, s)
 	r = make([]fr.Element, s)
 	o = make([]fr.Element, s)
-	partialL = make([]fr.Element, s)
 
 	for i := 0; i < spr.NbPublicVariables; i++ { // placeholders
 		l[i].Set(&solution[i])
@@ -72,28 +70,25 @@ func ComputeLRO(spr *cs.SparseR1CS, publicData *PublicRaw, solution []fr.Element
 		l[offset+i].Set(&solution[spr.Constraints[i].L.VariableID()])
 		r[offset+i].Set(&solution[spr.Constraints[i].R.VariableID()])
 		o[offset+i].Set(&solution[spr.Constraints[i].O.VariableID()])
-		partialL[offset+i].Set(&l[offset+i])
 	}
 	offset += len(spr.Constraints)
 	for i := 0; i < len(spr.Assertions); i++ { // assertions
 		l[offset+i].Set(&solution[spr.Assertions[i].L.VariableID()])
 		r[offset+i].Set(&solution[spr.Assertions[i].R.VariableID()])
 		o[offset+i].Set(&solution[spr.Assertions[i].O.VariableID()])
-		partialL[offset+i].Set(&l[offset+i])
 	}
 	offset += len(spr.Assertions)
 	for i := 0; i < s-offset; i++ { // offset to reach 2**n constraints (where the id of l,r,o is 0, so we assign solution[0])
 		l[offset+i].Set(&solution[0])
 		r[offset+i].Set(&solution[0])
 		o[offset+i].Set(&solution[0])
-		partialL[offset+i].Set(&l[offset+i])
 	}
 
-	return l, r, o, partialL
+	return l, r, o
 
 }
 
-// ComputeZ computes Z (in Lagrange basis), where:
+// ComputeZBis computes Z (in Lagrange basis), where:
 //
 // * Z of degree n (domainNum.Cardinality)
 // * Z(1)=1
@@ -102,17 +97,17 @@ func ComputeLRO(spr *cs.SparseR1CS, publicData *PublicRaw, solution []fr.Element
 //								     (l_i+s1+gamma)*(r_i+s2+gamma)*(o_i+s3+gamma)
 //
 //	* l, r, o are the solution in Lagrange basis
-func ComputeZ(l, r, o bn254.Polynomial, publicData *PublicRaw, gamma fr.Element) bn254.Polynomial {
+func ComputeZBis(l, r, o bn254.Polynomial, pk *ProvingKey, gamma fr.Element) bn254.Polynomial {
 
-	z := make(bn254.Polynomial, publicData.DomainNum.Cardinality)
-	nbElmts := int(publicData.DomainNum.Cardinality)
+	z := make(bn254.Polynomial, pk.DomainNum.Cardinality)
+	nbElmts := int(pk.DomainNum.Cardinality)
 
 	var f [3]fr.Element
 	var g [3]fr.Element
 	var u [3]fr.Element
 	u[0].SetOne()
-	u[1].Set(&publicData.Shifter[0])
-	u[2].Set(&publicData.Shifter[1])
+	u[1].Set(&pk.Vk.Shifter[0])
+	u[2].Set(&pk.Vk.Shifter[1])
 
 	z[0].SetOne()
 
@@ -122,46 +117,46 @@ func ComputeZ(l, r, o bn254.Polynomial, publicData *PublicRaw, gamma fr.Element)
 		f[1].Add(&r[i], &u[1]).Add(&f[1], &gamma) //r_i+u*z**i+gamma
 		f[2].Add(&o[i], &u[2]).Add(&f[2], &gamma) //o_i+u**2*z**i+gamma
 
-		g[0].Add(&l[i], &publicData.LS1[i]).Add(&g[0], &gamma) //l_i+z**i+gamma
-		g[1].Add(&r[i], &publicData.LS2[i]).Add(&g[1], &gamma) //r_i+u*z**i+gamma
-		g[2].Add(&o[i], &publicData.LS3[i]).Add(&g[2], &gamma) //o_i+u**2*z**i+gamma
+		g[0].Add(&l[i], &pk.LS1[i]).Add(&g[0], &gamma) //l_i+z**i+gamma
+		g[1].Add(&r[i], &pk.LS2[i]).Add(&g[1], &gamma) //r_i+u*z**i+gamma
+		g[2].Add(&o[i], &pk.LS3[i]).Add(&g[2], &gamma) //o_i+u**2*z**i+gamma
 
 		f[0].Mul(&f[0], &f[1]).Mul(&f[0], &f[2]) // (l_i+z**i+gamma)*(r_i+u*z**i+gamma)*(o_i+u**2z**i+gamma)
 		g[0].Mul(&g[0], &g[1]).Mul(&g[0], &g[2]) //  (l_i+s1+gamma)*(r_i+s2+gamma)*(o_i+s3+gamma)
 
 		z[i+1].Mul(&z[i], &f[0]).Div(&z[i+1], &g[0])
 
-		u[0].Mul(&u[0], &publicData.DomainNum.Generator) // z**i -> z**i+1
-		u[1].Mul(&u[1], &publicData.DomainNum.Generator) // u*z**i -> u*z**i+1
-		u[2].Mul(&u[2], &publicData.DomainNum.Generator) // u**2*z**i -> u**2*z**i+1
+		u[0].Mul(&u[0], &pk.DomainNum.Generator) // z**i -> z**i+1
+		u[1].Mul(&u[1], &pk.DomainNum.Generator) // u*z**i -> u*z**i+1
+		u[2].Mul(&u[2], &pk.DomainNum.Generator) // u**2*z**i -> u**2*z**i+1
 	}
 
 	return z
 
 }
 
-// evalConstraints computes the evaluation of lL+qrR+qqmL.R+qoO+k on
+// evalConstraintsBis computes the evaluation of lL+qrR+qqmL.R+qoO+k on
 // the odd cosets of (Z/8mZ)/(Z/mZ), where m=nbConstraints+nbAssertions.
-func evalConstraints(publicData *PublicRaw, evalL, evalR, evalO []fr.Element) []fr.Element {
+func evalConstraintsBis(pk *ProvingKey, evalL, evalR, evalO, qk []fr.Element) []fr.Element {
 
-	res := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
+	res := make([]fr.Element, 4*pk.DomainNum.Cardinality)
 
 	// evaluates ql, qr, qm, qo, k on the odd cosets of (Z/8mZ)/(Z/mZ)
-	evalQl := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evalQr := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evalQm := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evalQo := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evalQk := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evaluateCosets(publicData.Ql, evalQl, publicData.DomainNum)
-	evaluateCosets(publicData.Qr, evalQr, publicData.DomainNum)
-	evaluateCosets(publicData.Qm, evalQm, publicData.DomainNum)
-	evaluateCosets(publicData.Qo, evalQo, publicData.DomainNum)
-	evaluateCosets(publicData.Qk, evalQk, publicData.DomainNum)
+	evalQl := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evalQr := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evalQm := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evalQo := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evalQk := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evaluateCosetsBis(pk.Ql, evalQl, pk.DomainNum)
+	evaluateCosetsBis(pk.Qr, evalQr, pk.DomainNum)
+	evaluateCosetsBis(pk.Qm, evalQm, pk.DomainNum)
+	evaluateCosetsBis(pk.Qo, evalQo, pk.DomainNum)
+	evaluateCosetsBis(qk, evalQk, pk.DomainNum)
 
 	// computes the evaluation of qrR+qlL+qmL.R+qoO+k on the odd cosets
 	// of (Z/8mZ)/(Z/mZ)
 	var acc, buf fr.Element
-	for i := uint64(0); i < 4*publicData.DomainNum.Cardinality; i++ {
+	for i := uint64(0); i < 4*pk.DomainNum.Cardinality; i++ {
 
 		acc.Mul(&evalQl[i], &evalL[i]) // ql.l
 
@@ -179,18 +174,18 @@ func evalConstraints(publicData *PublicRaw, evalL, evalR, evalO []fr.Element) []
 	return res
 }
 
-// evalIDCosets id, uid, u**2id on the odd cosets of (Z/8mZ)/(Z/mZ)
-func evalIDCosets(publicData *PublicRaw) (id, uid, uuid bn254.Polynomial) {
+// evalIDCosetsBis id, uid, u**2id on the odd cosets of (Z/8mZ)/(Z/mZ)
+func evalIDCosetsBis(pk *ProvingKey) (id, uid, uuid bn254.Polynomial) {
 
 	// evaluation of id, uid, u**id on the cosets
-	id = make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	c := int(publicData.DomainNum.Cardinality)
+	id = make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	c := int(pk.DomainNum.Cardinality)
 	id[0].SetOne()
 	id[1].SetOne()
 	id[2].SetOne()
 	id[3].SetOne()
 	for i := 1; i < c; i++ {
-		id[4*i].Mul(&id[4*(i-1)], &publicData.DomainNum.Generator)
+		id[4*i].Mul(&id[4*(i-1)], &pk.DomainNum.Generator)
 		id[4*i+1].Set(&id[4*i])
 		id[4*i+2].Set(&id[4*i])
 		id[4*i+3].Set(&id[4*i])
@@ -198,14 +193,14 @@ func evalIDCosets(publicData *PublicRaw) (id, uid, uuid bn254.Polynomial) {
 	// at this stage, id = [1,1,1,1,|z,z,z,z|,...,|z**n-1,z**n-1,z**n-1,z**n-1]
 
 	var uu fr.Element
-	uu.Square(&publicData.DomainNum.FinerGenerator)
+	uu.Square(&pk.DomainNum.FinerGenerator)
 	var u [4]fr.Element
-	u[0].Set(&publicData.DomainNum.FinerGenerator)                // u
-	u[1].Mul(&u[0], &uu)                                          // u**3
-	u[2].Mul(&u[1], &uu)                                          // u**5
-	u[3].Mul(&u[2], &uu)                                          // u**7
-	uid = make([]fr.Element, 4*publicData.DomainNum.Cardinality)  // shifter[0]*ID evaluated on odd cosets of (Z/8mZ)/(Z/mZ)
-	uuid = make([]fr.Element, 4*publicData.DomainNum.Cardinality) // shifter[1]*ID evaluated on odd cosets of (Z/8mZ)/(Z/mZ)
+	u[0].Set(&pk.DomainNum.FinerGenerator)                // u
+	u[1].Mul(&u[0], &uu)                                  // u**3
+	u[2].Mul(&u[1], &uu)                                  // u**5
+	u[3].Mul(&u[2], &uu)                                  // u**7
+	uid = make([]fr.Element, 4*pk.DomainNum.Cardinality)  // shifter[0]*ID evaluated on odd cosets of (Z/8mZ)/(Z/mZ)
+	uuid = make([]fr.Element, 4*pk.DomainNum.Cardinality) // shifter[1]*ID evaluated on odd cosets of (Z/8mZ)/(Z/mZ)
 	for i := 0; i < c; i++ {
 
 		id[4*i].Mul(&id[4*i], &u[0])     // coset u.<1,z,..,z**n-1>
@@ -213,44 +208,44 @@ func evalIDCosets(publicData *PublicRaw) (id, uid, uuid bn254.Polynomial) {
 		id[4*i+2].Mul(&id[4*i+2], &u[2]) // coset u**5.<1,z,..,z**n-1>
 		id[4*i+3].Mul(&id[4*i+3], &u[3]) // coset u**7.<1,z,..,z**n-1>
 
-		uid[4*i].Mul(&id[4*i], &publicData.Shifter[0])     // shifter[0]*ID
-		uid[4*i+1].Mul(&id[4*i+1], &publicData.Shifter[0]) // shifter[0]*ID
-		uid[4*i+2].Mul(&id[4*i+2], &publicData.Shifter[0]) // shifter[0]*ID
-		uid[4*i+3].Mul(&id[4*i+3], &publicData.Shifter[0]) // shifter[0]*ID
+		uid[4*i].Mul(&id[4*i], &pk.Vk.Shifter[0])     // shifter[0]*ID
+		uid[4*i+1].Mul(&id[4*i+1], &pk.Vk.Shifter[0]) // shifter[0]*ID
+		uid[4*i+2].Mul(&id[4*i+2], &pk.Vk.Shifter[0]) // shifter[0]*ID
+		uid[4*i+3].Mul(&id[4*i+3], &pk.Vk.Shifter[0]) // shifter[0]*ID
 
-		uuid[4*i].Mul(&id[4*i], &publicData.Shifter[1])     // shifter[1]*ID
-		uuid[4*i+1].Mul(&id[4*i+1], &publicData.Shifter[1]) // shifter[1]*ID
-		uuid[4*i+2].Mul(&id[4*i+2], &publicData.Shifter[1]) // shifter[1]*ID
-		uuid[4*i+3].Mul(&id[4*i+3], &publicData.Shifter[1]) // shifter[1]*ID
+		uuid[4*i].Mul(&id[4*i], &pk.Vk.Shifter[1])     // shifter[1]*ID
+		uuid[4*i+1].Mul(&id[4*i+1], &pk.Vk.Shifter[1]) // shifter[1]*ID
+		uuid[4*i+2].Mul(&id[4*i+2], &pk.Vk.Shifter[1]) // shifter[1]*ID
+		uuid[4*i+3].Mul(&id[4*i+3], &pk.Vk.Shifter[1]) // shifter[1]*ID
 
 	}
 	return
 }
 
-// evalConstraintOrdering computes the evaluation of Z(uX)g1g2g3-Z(X)f1f2f3 on the odd
+// evalConstraintOrderingBis computes the evaluation of Z(uX)g1g2g3-Z(X)f1f2f3 on the odd
 // cosets of (Z/8mZ)/(Z/mZ), where m=nbConstraints+nbAssertions.
 //
 // z: permutation accumulator polynomial in canonical form
 // l, r, o: solution, in canonical form
-func evalConstraintOrdering(publicData *PublicRaw, evalZ, evalZu, evalL, evalR, evalO bn254.Polynomial, gamma fr.Element) bn254.Polynomial {
+func evalConstraintOrderingBis(pk *ProvingKey, evalZ, evalZu, evalL, evalR, evalO bn254.Polynomial, gamma fr.Element) bn254.Polynomial {
 
 	// evaluation of z, zu, s1, s2, s3, on the odd cosets of (Z/8mZ)/(Z/mZ)
-	evalS1 := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evalS2 := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evalS3 := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evaluateCosets(publicData.CS1, evalS1, publicData.DomainNum)
-	evaluateCosets(publicData.CS2, evalS2, publicData.DomainNum)
-	evaluateCosets(publicData.CS3, evalS3, publicData.DomainNum)
+	evalS1 := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evalS2 := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evalS3 := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evaluateCosetsBis(pk.CS1, evalS1, pk.DomainNum)
+	evaluateCosetsBis(pk.CS2, evalS2, pk.DomainNum)
+	evaluateCosetsBis(pk.CS3, evalS3, pk.DomainNum)
 
 	// evalutation of ID, u*ID, u**2*ID on the odd cosets of (Z/8mZ)/(Z/mZ)
-	evalID, evaluID, evaluuID := evalIDCosets(publicData)
+	evalID, evaluID, evaluuID := evalIDCosetsBis(pk)
 
 	// computes Z(uX)g1g2g3l-Z(X)f1f2f3l on the odd cosets of (Z/8mZ)/(Z/mZ)
-	res := make(bn254.Polynomial, 4*publicData.DomainNum.Cardinality)
+	res := make(bn254.Polynomial, 4*pk.DomainNum.Cardinality)
 
 	var f [3]fr.Element
 	var g [3]fr.Element
-	for i := 0; i < 4*int(publicData.DomainNum.Cardinality); i++ {
+	for i := 0; i < 4*int(pk.DomainNum.Cardinality); i++ {
 
 		f[0].Add(&evalL[i], &evalID[i]).Add(&f[0], &gamma)   //l_i+z**i+gamma
 		f[1].Add(&evalR[i], &evaluID[i]).Add(&f[1], &gamma)  //r_i+u*z**i+gamma
@@ -274,26 +269,26 @@ func evalConstraintOrdering(publicData *PublicRaw, evalZ, evalZu, evalL, evalR, 
 	return res
 }
 
-// evalStartsAtOne computes the evaluation of L1*(z-1) on the odd cosets
+// evalStartsAtOneBis computes the evaluation of L1*(z-1) on the odd cosets
 // of (Z/8mZ)/(Z/mZ).
 //
 // evalZ is the evaluation of z (=permutation constraint polynomial) on odd cosets of (Z/8mZ)/(Z/mZ)
-func evalStartsAtOne(publicData *PublicRaw, evalZ bn254.Polynomial) bn254.Polynomial {
+func evalStartsAtOneBis(pk *ProvingKey, evalZ bn254.Polynomial) bn254.Polynomial {
 
 	// computes L1 (canonical form)
-	lOneLagrange := make([]fr.Element, publicData.DomainNum.Cardinality)
+	lOneLagrange := make([]fr.Element, pk.DomainNum.Cardinality)
 	lOneLagrange[0].SetOne()
-	publicData.DomainNum.FFTInverse(lOneLagrange, fft.DIF, 0)
+	pk.DomainNum.FFTInverse(lOneLagrange, fft.DIF, 0)
 	fft.BitReverse(lOneLagrange)
 
 	// evaluates L1 on the odd cosets of (Z/8mZ)/(Z/mZ)
-	res := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evaluateCosets(lOneLagrange, res, publicData.DomainNum)
+	res := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evaluateCosetsBis(lOneLagrange, res, pk.DomainNum)
 
 	// // evaluates L1*(z-1) on the odd cosets of (Z/8mZ)/(Z/mZ)
 	var buf, one fr.Element
 	one.SetOne()
-	for i := 0; i < 4*int(publicData.DomainNum.Cardinality); i++ {
+	for i := 0; i < 4*int(pk.DomainNum.Cardinality); i++ {
 		buf.Sub(&evalZ[i], &one)
 		res[i].Mul(&buf, &res[i])
 	}
@@ -301,14 +296,14 @@ func evalStartsAtOne(publicData *PublicRaw, evalZ bn254.Polynomial) bn254.Polyno
 	return res
 }
 
-// evaluateCosets evaluates poly (canonical form) of degree m=domainNum.Cardinality on
+// evaluateCosetsBis evaluates poly (canonical form) of degree m=domainNum.Cardinality on
 // the 4 odd cosets of (Z/8mZ)/(Z/mZ), so it dodges Z/mZ (+Z/2kmZ), which contains the
 // vanishing set of Z.
 //
 // Puts the result in res (of size 4*domain.Cardinality).
 //
 // Both sizes of poly and res are powers of 2, len(res) = 4*len(poly).
-func evaluateCosets(poly, res []fr.Element, domain *fft.Domain) {
+func evaluateCosetsBis(poly, res []fr.Element, domain *fft.Domain) {
 
 	// build a copy of poly padded with 0 so it has the length of the closest power of 2 of poly
 	evaluations := make([][]fr.Element, 4)
@@ -340,8 +335,8 @@ func evaluateCosets(poly, res []fr.Element, domain *fft.Domain) {
 	}
 }
 
-// shiftZ turns z to z(uX) (both in Lagrange basis)
-func shiftZ(z bn254.Polynomial) bn254.Polynomial {
+// shitZBis turns z to z(uX) (both in Lagrange basis)
+func shitZBis(z bn254.Polynomial) bn254.Polynomial {
 
 	res := make(bn254.Polynomial, len(z))
 	copy(res, z)
@@ -356,26 +351,26 @@ func shiftZ(z bn254.Polynomial) bn254.Polynomial {
 	return res
 }
 
-// computeH computes h in canonical form, split as h1+X^mh2+X^2mh3 such that
+// computeHBis computes h in canonical form, split as h1+X^mh2+X^2mh3 such that
 //
 // qlL+qrR+qmL.R+qoO+k + alpha.(zu*g1*g2*g3*l-z*f1*f2*f3*l) + alpha**2*L1*(z-1)= h.Z
 // \------------------/         \------------------------/             \-----/
 //    constraintsInd			    constraintOrdering					startsAtOne
 //
 // constraintInd, constraintOrdering are evaluated on the odd cosets of (Z/8mZ)/(Z/mZ)
-func computeH(publicData *PublicRaw, constraintsInd, constraintOrdering, startsAtOne bn254.Polynomial, alpha fr.Element) (bn254.Polynomial, bn254.Polynomial, bn254.Polynomial) {
+func computeHBis(pk *ProvingKey, constraintsInd, constraintOrdering, startsAtOne bn254.Polynomial, alpha fr.Element) (bn254.Polynomial, bn254.Polynomial, bn254.Polynomial) {
 
-	h := make(bn254.Polynomial, publicData.DomainH.Cardinality)
+	h := make(bn254.Polynomial, pk.DomainH.Cardinality)
 
 	// evaluate Z = X**m-1 on the odd cosets of (Z/8mZ)/(Z/mZ)
 	var bExpo big.Int
-	bExpo.SetUint64(publicData.DomainNum.Cardinality)
+	bExpo.SetUint64(pk.DomainNum.Cardinality)
 	var u [4]fr.Element
 	var uu fr.Element
 	var one fr.Element
 	one.SetOne()
-	uu.Square(&publicData.DomainNum.FinerGenerator)
-	u[0].Set(&publicData.DomainNum.FinerGenerator)
+	uu.Square(&pk.DomainNum.FinerGenerator)
+	u[0].Set(&pk.DomainNum.FinerGenerator)
 	u[1].Mul(&u[0], &uu)
 	u[2].Mul(&u[1], &uu)
 	u[3].Mul(&u[2], &uu)
@@ -386,7 +381,7 @@ func computeH(publicData *PublicRaw, constraintsInd, constraintOrdering, startsA
 
 	// evaluate qlL+qrR+qmL.R+qoO+k + alpha.(zu*g1*g2*g3*l-z*f1*f2*f3*l) + alpha**2*L1(X)(Z(X)-1)
 	// on the odd cosets of (Z/8mZ)/(Z/mZ)
-	for i := 0; i < 4*int(publicData.DomainNum.Cardinality); i++ {
+	for i := 0; i < 4*int(pk.DomainNum.Cardinality); i++ {
 		h[i].Mul(&startsAtOne[i], &alpha).
 			Add(&h[i], &constraintOrdering[i]).
 			Mul(&h[i], &alpha).
@@ -395,7 +390,7 @@ func computeH(publicData *PublicRaw, constraintsInd, constraintOrdering, startsA
 
 	// evaluate qlL+qrR+qmL.R+qoO+k + alpha.(zu*g1*g2*g3*l-z*f1*f2*f3*l)/Z
 	// on the odd cosets of (Z/8mZ)/(Z/mZ)
-	for i := 0; i < int(publicData.DomainNum.Cardinality); i++ {
+	for i := 0; i < int(pk.DomainNum.Cardinality); i++ {
 		h[4*i].Mul(&h[4*i], &u[0])
 		h[4*i+1].Mul(&h[4*i+1], &u[1])
 		h[4*i+2].Mul(&h[4*i+2], &u[2])
@@ -403,206 +398,325 @@ func computeH(publicData *PublicRaw, constraintsInd, constraintOrdering, startsA
 	}
 
 	// put h in canonical form
-	publicData.DomainH.FFTInverse(h, fft.DIF, 1)
+	pk.DomainH.FFTInverse(h, fft.DIF, 1)
 	fft.BitReverse(h)
 
-	h1 := make(bn254.Polynomial, publicData.DomainNum.Cardinality)
-	h2 := make(bn254.Polynomial, publicData.DomainNum.Cardinality)
-	h3 := make(bn254.Polynomial, publicData.DomainNum.Cardinality)
-	copy(h1, h[:publicData.DomainNum.Cardinality])
-	copy(h2, h[publicData.DomainNum.Cardinality:2*publicData.DomainNum.Cardinality])
-	copy(h3, h[2*publicData.DomainNum.Cardinality:3*publicData.DomainNum.Cardinality])
+	h1 := make(bn254.Polynomial, pk.DomainNum.Cardinality)
+	h2 := make(bn254.Polynomial, pk.DomainNum.Cardinality)
+	h3 := make(bn254.Polynomial, pk.DomainNum.Cardinality)
+	copy(h1, h[:pk.DomainNum.Cardinality])
+	copy(h2, h[pk.DomainNum.Cardinality:2*pk.DomainNum.Cardinality])
+	copy(h3, h[2*pk.DomainNum.Cardinality:3*pk.DomainNum.Cardinality])
 
 	return h1, h2, h3
 
 }
 
-// ProveRaw from the public data
-// TODO add a parameter to force the resolution of the system even if a constraint does not hold
-func ProveRaw(spr *cs.SparseR1CS, publicData *PublicRaw, fullWitness bn254witness.Witness) (*ProofRaw, error) {
+// computeLinearizedPolynomial computes the linearized polynomial in canonical basis.
+// The purpose is to commit and open all in one ql, qr, qm, qo, qk.
+// * a, b, c are the evaluation of l, r, o at zeta
+// * z is the permutation polynomial, zu is Z(uX), the shifted version of Z
+// * pk is the proving key: the linearized polynomial is a linear combination of ql, qr, qm, qo, qk.
+func computeLinearizedPolynomial(l, r, o, alpha, gamma, zeta, zu fr.Element, z bn254.Polynomial, pk *ProvingKey) bn254.Polynomial {
+
+	// first part: individual constraints
+	var rl fr.Element
+	rl.Mul(&r, &l)
+	_linearizedPolynomial := pk.Qm.Clone()
+	_linearizedPolynomial.ScaleInPlace(rl) // linPol = lr*Qm
+
+	tmp := pk.Ql.Clone()
+	tmp.ScaleInPlace(l)
+	_linearizedPolynomial.Add(_linearizedPolynomial, tmp) // linPol = lr*Qm + l*Ql
+
+	tmp = pk.Qr.Clone()
+	tmp.ScaleInPlace(r)
+	_linearizedPolynomial.Add(_linearizedPolynomial, tmp) // linPol = lr*Qm + l*Ql + r*Qr
+
+	tmp = pk.Qo.Clone()
+	tmp.ScaleInPlace(o)
+	_linearizedPolynomial.Add(_linearizedPolynomial, tmp) // linPol = lr*Qm + l*Ql + r*Qr + o*Qo
+
+	_linearizedPolynomial.Add(_linearizedPolynomial, &pk.CQk) // linPol = lr*Qm + l*Ql + r*Qr + o*Qo + Qk
+
+	// second part: Z(uzeta)(a+s1+gamma)*(b+s2+gamma)*s3(X)-Z(X)(a+zeta+gamma)*(b+uzeta+gamma)*(c+u**2*zeta+gamma)
+	var s1, s2, t fr.Element
+	s1.SetInterface(pk.CS1.Eval(zeta)).Add(&s1, &l).Add(&s1, &gamma) // (a+s1+gamma)
+	t.SetInterface(pk.CS2.Eval(zeta)).Add(&t, &r).Add(&t, &gamma)    // (b+s2+gamma)
+	s1.Mul(&s1, &t).                                                 // (a+s1+gamma)*(b+s2+gamma)
+										Mul(&s1, &zu) // (a+s1+gamma)*(b+s2+gamma)*Z(uzeta)
+
+	s2.Add(&l, &zeta).Add(&s2, &gamma)                          // (a+z+gamma)
+	t.Mul(&pk.Vk.Shifter[0], &zeta).Add(&t, &r).Add(&t, &gamma) // (b+uz+gamma)
+	s2.Mul(&s2, &t)                                             // (a+z+gamma)*(b+uz+gamma)
+	t.Mul(&pk.Vk.Shifter[1], &zeta).Add(&t, &o).Add(&t, &gamma) // (o+u**2z+gamma)
+	s2.Mul(&s2, &t)                                             // (a+z+gamma)*(b+uz+gamma)*(c+u**2*z+gamma)
+	s2.Neg(&s2)                                                 // -(a+z+gamma)*(b+uz+gamma)*(c+u**2*z+gamma)
+
+	p1 := pk.CS3.Clone()
+	p1.ScaleInPlace(s1) // (a+s1+gamma)*(b+s2+gamma)*Z(uzeta)*s3(X)
+	p2 := z.Clone()
+	p2.ScaleInPlace(s2) // -Z(X)(a+zeta+gamma)*(b+uzeta+gamma)*(c+u**2*zeta+gamma)
+	p1.Add(p1, p2)
+	p1.ScaleInPlace(alpha) // alpha*( Z(uzeta)*(a+s1+gamma)*(b+s2+gamma)s3(X)-Z(X)(a+zeta+gamma)*(b+uzeta+gamma)*(c+u**2*zeta+gamma) )
+
+	_linearizedPolynomial.Add(_linearizedPolynomial, p1)
+
+	// third part L1(zeta)*alpha**2**Z
+	var lagrange, one, den, frNbElmt fr.Element
+	one.SetOne()
+	nbElmt := int64(pk.DomainNum.Cardinality)
+	lagrange.Set(&zeta).
+		Exp(lagrange, big.NewInt(nbElmt)).
+		Sub(&lagrange, &one)
+	frNbElmt.SetUint64(uint64(nbElmt))
+	den.Sub(&zeta, &one).
+		Mul(&den, &frNbElmt).
+		Inverse(&den)
+	lagrange.Mul(&lagrange, &den). // L_0 = 1/m*(zeta**n-1)/(zeta-1)
+					Mul(&lagrange, &alpha).
+					Mul(&lagrange, &alpha) // alpha**2*L_0
+	p1 = z.Clone()
+	p1.ScaleInPlace(lagrange)
+
+	// finish the computation
+	_linearizedPolynomial.Add(_linearizedPolynomial, p1)
+
+	linearizedPolynomial := _linearizedPolynomial.(*bn254.Polynomial)
+
+	return *linearizedPolynomial
+}
+
+// Prove from the public data
+func Prove(spr *cs.SparseR1CS, pk *ProvingKey, fullWitness bn254witness.Witness) (*Proof, error) {
 
 	// create a transcript manager to apply Fiat Shamir
 	fs := fiatshamir.NewTranscript(fiatshamir.SHA256, "gamma", "alpha", "zeta")
 
 	// result
-	proof := &ProofRaw{}
+	proofBis := &Proof{}
 
 	// compute the solution
-	solution, _ := spr.Solve(fullWitness)
+	solution, err := spr.Solve(fullWitness)
+	if err != nil {
+		return proofBis, err
+	}
 
 	// query l, r, o in Lagrange basis
-	ll, lr, lo, partialL := ComputeLRO(spr, publicData, solution)
+	ll, lr, lo := ComputeLROBis(spr, pk, solution)
 
 	// save ll, lr, lo, and make a copy of them in canonical basis.
 	// We commit them and derive gamma from them.
-	cl := make(bn254.Polynomial, len(ll))
-	cr := make(bn254.Polynomial, len(lr))
-	co := make(bn254.Polynomial, len(lo))
+	sizeCommon := int64(pk.DomainNum.Cardinality)
+	cl := make(bn254.Polynomial, sizeCommon)
+	cr := make(bn254.Polynomial, sizeCommon)
+	co := make(bn254.Polynomial, sizeCommon)
 	copy(cl, ll)
 	copy(cr, lr)
 	copy(co, lo)
-	publicData.DomainNum.FFTInverse(cl, fft.DIF, 0)
-	publicData.DomainNum.FFTInverse(cr, fft.DIF, 0)
-	publicData.DomainNum.FFTInverse(co, fft.DIF, 0)
-	publicData.DomainNum.FFTInverse(partialL, fft.DIF, 0)
+	pk.DomainNum.FFTInverse(cl, fft.DIF, 0)
+	pk.DomainNum.FFTInverse(cr, fft.DIF, 0)
+	pk.DomainNum.FFTInverse(co, fft.DIF, 0)
 	fft.BitReverse(cl)
 	fft.BitReverse(cr)
 	fft.BitReverse(co)
-	fft.BitReverse(partialL)
 
 	// derive gamma from the Comm(l), Comm(r), Comm(o)
-	var err error
-	proof.CommitmentsLROZH[0], err = publicData.CommitmentScheme.Commit(&cl)
+	proofBis.LRO[0], _ = pk.Vk.CommitmentScheme.Commit(&cl)
+	proofBis.LRO[1], _ = pk.Vk.CommitmentScheme.Commit(&cr)
+	proofBis.LRO[2], _ = pk.Vk.CommitmentScheme.Commit(&co)
+	err = fs.Bind("gamma", proofBis.LRO[0].Marshal())
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
-	proof.CommitmentsLROZH[1], err = publicData.CommitmentScheme.Commit(&cr)
+	err = fs.Bind("gamma", proofBis.LRO[1].Marshal())
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
-	proof.CommitmentsLROZH[2], err = publicData.CommitmentScheme.Commit(&co)
+	err = fs.Bind("gamma", proofBis.LRO[2].Marshal())
 	if err != nil {
-		return proof, err
-	}
-	err = fs.Bind("gamma", proof.CommitmentsLROZH[0].Marshal())
-	if err != nil {
-		return proof, err
-	}
-	err = fs.Bind("gamma", proof.CommitmentsLROZH[1].Marshal())
-	if err != nil {
-		return proof, err
-	}
-	err = fs.Bind("gamma", proof.CommitmentsLROZH[2].Marshal())
-	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
 	bgamma, err := fs.ComputeChallenge("gamma")
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
 	var gamma fr.Element
 	gamma.SetBytes(bgamma)
 
 	// compute Z, the permutation accumulator polynomial, in Lagrange basis
-	z := ComputeZ(ll, lr, lo, publicData, gamma)
+	z := ComputeZBis(ll, lr, lo, pk, gamma)
 
 	// compute Z(uX), in Lagrange basis
-	zu := shiftZ(z)
+	zu := shitZBis(z)
 
 	// compute the evaluations of l, r, o on odd cosets of (Z/8mZ)/(Z/mZ)
-	evalL := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evalR := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evalO := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evaluateCosets(cl, evalL, publicData.DomainNum)
-	evaluateCosets(cr, evalR, publicData.DomainNum)
-	evaluateCosets(co, evalO, publicData.DomainNum)
+	evalL := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evalR := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evalO := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evaluateCosetsBis(cl, evalL, pk.DomainNum)
+	evaluateCosetsBis(cr, evalR, pk.DomainNum)
+	evaluateCosetsBis(co, evalO, pk.DomainNum)
+
+	// compute qk in canonical basis, completed with the public inputs
+	qkFullC := make(bn254.Polynomial, sizeCommon)
+	copy(qkFullC, fullWitness[:spr.NbPublicVariables])
+	copy(qkFullC[spr.NbPublicVariables:], pk.LQk[spr.NbPublicVariables:])
+	pk.DomainNum.FFTInverse(qkFullC, fft.DIF, 0)
+	fft.BitReverse(qkFullC)
 
 	// compute the evaluation of qlL+qrR+qmL.R+qoO+k on the odd cosets of (Z/8mZ)/(Z/mZ)
-	constraintsInd := evalConstraints(publicData, evalL, evalR, evalO)
+	constraintsInd := evalConstraintsBis(pk, evalL, evalR, evalO, qkFullC)
 
 	// put back z, zu in canonical basis
-	publicData.DomainNum.FFTInverse(z, fft.DIF, 0)
-	publicData.DomainNum.FFTInverse(zu, fft.DIF, 0)
+	pk.DomainNum.FFTInverse(z, fft.DIF, 0)
+	pk.DomainNum.FFTInverse(zu, fft.DIF, 0)
 	fft.BitReverse(z)
 	fft.BitReverse(zu)
 
 	// evaluate z, zu on the odd cosets of (Z/8mZ)/(Z/mZ)
-	evalZ := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evalZu := make([]fr.Element, 4*publicData.DomainNum.Cardinality)
-	evaluateCosets(z, evalZ, publicData.DomainNum)
-	evaluateCosets(zu, evalZu, publicData.DomainNum)
+	evalZ := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evalZu := make([]fr.Element, 4*pk.DomainNum.Cardinality)
+	evaluateCosetsBis(z, evalZ, pk.DomainNum)
+	evaluateCosetsBis(zu, evalZu, pk.DomainNum)
 
 	// compute zu*g1*g2*g3-z*f1*f2*f3 on the odd cosets of (Z/8mZ)/(Z/mZ)
-	constraintsOrdering := evalConstraintOrdering(publicData, evalZ, evalZu, evalL, evalR, evalO, gamma)
+	constraintsOrdering := evalConstraintOrderingBis(pk, evalZ, evalZu, evalL, evalR, evalO, gamma)
 
 	// compute L1*(z-1) on the odd cosets of (Z/8mZ)/(Z/mZ)
-	startsAtOne := evalStartsAtOne(publicData, evalZ)
+	startsAtOne := evalStartsAtOneBis(pk, evalZ)
 
 	// commit to Z
-	proof.CommitmentsLROZH[3], err = publicData.CommitmentScheme.Commit(&z)
+	proofBis.Z, err = pk.Vk.CommitmentScheme.Commit(&z)
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
 
 	// derive alpha from the Comm(l), Comm(r), Comm(o), Com(Z)
-	err = fs.Bind("alpha", proof.CommitmentsLROZH[3].Marshal())
+	err = fs.Bind("alpha", proofBis.Z.Marshal())
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
 	balpha, err := fs.ComputeChallenge("alpha")
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
 	var alpha fr.Element
 	alpha.SetBytes(balpha)
 
 	// compute h in canonical form
-	h1, h2, h3 := computeH(publicData, constraintsInd, constraintsOrdering, startsAtOne, alpha)
+	h1, h2, h3 := computeHBis(pk, constraintsInd, constraintsOrdering, startsAtOne, alpha)
 
 	// commit to h (3 commitments h1 + x**n*h2 + x**2n*h3)
-	proof.CommitmentsLROZH[4], err = publicData.CommitmentScheme.Commit(&h1)
+	proofBis.H[0], err = pk.Vk.CommitmentScheme.Commit(&h1)
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
-	proof.CommitmentsLROZH[5], err = publicData.CommitmentScheme.Commit(&h2)
+	proofBis.H[1], err = pk.Vk.CommitmentScheme.Commit(&h2)
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
-	proof.CommitmentsLROZH[6], err = publicData.CommitmentScheme.Commit(&h3)
+	proofBis.H[2], err = pk.Vk.CommitmentScheme.Commit(&h3)
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
 
 	// derive zeta, the point of evaluation
-	err = fs.Bind("zeta", proof.CommitmentsLROZH[4].Marshal())
+	err = fs.Bind("zeta", proofBis.H[0].Marshal())
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
-	err = fs.Bind("zeta", proof.CommitmentsLROZH[5].Marshal())
+	err = fs.Bind("zeta", proofBis.H[1].Marshal())
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
-	err = fs.Bind("zeta", proof.CommitmentsLROZH[6].Marshal())
+	err = fs.Bind("zeta", proofBis.H[2].Marshal())
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
 	bzeta, err := fs.ComputeChallenge("zeta")
 	if err != nil {
-		return proof, err
+		return proofBis, err
 	}
 	var zeta fr.Element
 	zeta.SetBytes(bzeta)
 
-	// compute the linearization polynomial r at zeta (goal: save committing separately to ql, qr, qm, qo, k)
+	// open Z at zeta*z
+	var zetaShifted fr.Element
+	zetaShifted.Mul(&zeta, &pk.Vk.Generator)
+	proofBis.ZShiftedOpening, _ = pk.Vk.CommitmentScheme.Open(
+		zetaShifted,
+		&z,
+	)
+	var zuzeta fr.Element
+	zuzeta.SetInterface(proofBis.ZShiftedOpening.GetClaimedValue())
 
 	// compute evaluations of l, r, o, z at zeta
-	proof.LROZH[0].SetInterface(partialL.Eval(&zeta).(fr.Element))
-	proof.LROZH[1].SetInterface(cr.Eval(&zeta).(fr.Element))
-	proof.LROZH[2].SetInterface(co.Eval(&zeta).(fr.Element))
-	proof.LROZH[3].SetInterface(z.Eval(&zeta).(fr.Element))
+	var lzeta, rzeta, ozeta fr.Element
+	lzeta.SetInterface(cl.Eval(&zeta))
+	rzeta.SetInterface(cr.Eval(&zeta))
+	ozeta.SetInterface(co.Eval(&zeta))
 
-	// compute evaluations of h1, h2, h3 at zeta (so h(zeta)=h1(zeta)+zeta^m*h2(zeta)+zeta^2m*h3(zeta))
-	proof.LROZH[4].SetInterface(h1.Eval(&zeta))
-	proof.LROZH[5].SetInterface(h2.Eval(&zeta))
-	proof.LROZH[6].SetInterface(h3.Eval(&zeta))
+	// compute the linearization polynomial r at zeta (goal: save committing separately to z, ql, qr, qm, qo, k)
+	linearizedPolynomial := computeLinearizedPolynomial(
+		lzeta,
+		rzeta,
+		ozeta,
+		alpha,
+		gamma,
+		zeta,
+		zuzeta,
+		z,
+		pk,
+	)
 
-	// compute evaluation of z at z*zeta
-	var zzeta fr.Element
-	zzeta.Mul(&zeta, &publicData.DomainNum.Generator)
-	proof.ZShift.SetInterface(z.Eval(&zzeta))
+	// foldedHDigest = Comm(h1) + zeta**m*Comm(h2) + zeta**2m*Comm(h3)
+	var bZetaPowerm big.Int
+	sizeBigInt := big.NewInt(sizeCommon)
+	var zetaPowerm fr.Element
+	zetaPowerm.Exp(zeta, sizeBigInt)
+	zetaPowerm.ToBigIntRegular(&bZetaPowerm)
+	foldedHDigest := proofBis.H[2].Clone()
+	foldedHDigest.ScalarMul(foldedHDigest, bZetaPowerm)
+	foldedHDigest.Add(foldedHDigest, proofBis.H[1])     // zeta**m*Comm(h3)
+	foldedHDigest.ScalarMul(foldedHDigest, bZetaPowerm) // zeta**2m*Comm(h3) + zeta**m*Comm(h2)
+	foldedHDigest.Add(foldedHDigest, proofBis.H[0])     // zeta**2m*Comm(h3) + zeta**m*Comm(h2) + Comm(h1)
 
-	// compute batfch opening proof for l, r, o, h, z at zeta
-	polynomialsToOpenAtZeta := []polynomial.Polynomial{&cl, &cr, &co, &z, &h1, &h2, &h3}
-	proof.BatchOpenings, err = publicData.CommitmentScheme.BatchOpenSinglePoint(&zeta, proof.CommitmentsLROZH[:], polynomialsToOpenAtZeta)
-	if err != nil {
-		return proof, err
-	}
+	// foldedH = h1 + zeta*h2 + zeta**2*h3
+	foldedH := h3.Clone()
+	foldedH.ScaleInPlace(bZetaPowerm) // zeta**m*h3
+	foldedH.Add(foldedH, &h2)         // zeta**m*h3+h2
+	foldedH.ScaleInPlace(bZetaPowerm) // zeta**2m*h3+h2*zeta**m
+	foldedH.Add(foldedH, &h1)         // zeta**2m*h3+zeta**m*h2 + h1
+	// foldedH correct
 
-	// compute opening proof for z at z*zeta
-	proof.OpeningZShift, err = publicData.CommitmentScheme.Open(&zzeta, &z)
-	if err != nil {
-		return proof, err
-	}
+	// TODO this commitment is only necessary to derive the challenge, we should
+	// be able to avoid doing it and get the challenge in another way
+	linearizedPolynomialDigest, _ := pk.Vk.CommitmentScheme.Commit(&linearizedPolynomial)
 
-	return proof, nil
+	// Batch open the first list of polynomials
+	proofBis.BatchedProof, _ = pk.Vk.CommitmentScheme.BatchOpenSinglePoint(
+		zeta,
+		[]polynomial.Digest{
+			foldedHDigest,
+			linearizedPolynomialDigest,
+			proofBis.LRO[0],
+			proofBis.LRO[1],
+			proofBis.LRO[2],
+			pk.Vk.S[0],
+			pk.Vk.S[1],
+		},
+		[]polynomial.Polynomial{
+			foldedH,
+			&linearizedPolynomial,
+			&cl,
+			&cr,
+			&co,
+			&pk.CS1,
+			&pk.CS2,
+		},
+	)
+
+	return proofBis, nil
+
 }
